Rename : TripFriends 패키지로 이동
